+++ medinify/process/__init__.py
@@ -1,5 +1,9 @@
 from .processor import Processor
 from .bow_processor import BowProcessor
-from .embeddings_processor import EmbeddingsProcessor, find_embeddings
+from .embeddings_processor import EmbeddingsProcessor
 from .matrix_processor import MatrixProcessor
 from .pos_processor import PosProcessor
+from .utils import find_embeddings
+from .utils import get_lookup_table
+from .utils import get_pos_list
+
