+++ medinify/process/embeddings_processor.py
@@ -1,9 +1,9 @@
 
-from medinify.process import *
+from medinify.process import Processor
+from medinify.process.utils import find_embeddings
 import numpy as np
 from gensim.models import KeyedVectors
 import warnings
-import os
 
 warnings.filterwarnings("ignore")
 
@@ -46,34 +46,3 @@ class EmbeddingsProcessor(Processor):
                 embeddings[i] = np.average(all_embeddings, axis=0)
         return embeddings
 
-
-def find_embeddings():
-    """
-    Searches of pretrained embeddings file in medinify/data/embeddings folder
-    :return: abspath (str) absolute path to embeddings file or None if not found
-    """
-    abspath = None
-    for file in os.walk(os.getcwd()):
-        if os.path.isdir(file[0]) and file[0][-24:] == 'medinify/data/embeddings':
-            directory_path = file[0]
-            embeddings_files = os.listdir(directory_path)
-            if not embeddings_files:
-                raise FileNotFoundError(
-                    'No word embeddings found at data/embeddings.')
-            elif len(embeddings_files) > 1:
-                print('Multiple embedding files found.\n'
-                      'Please specify which file to use (enter file name):')
-                while True:
-                    for filename in embeddings_files:
-                        print('\t%s' % filename)
-                    chosen_file = input()
-                    if chosen_file in embeddings_files:
-                        embeddings_file = chosen_file
-                        break
-                    else:
-                        print('Invalid file entered. '
-                              'Please specify which file to use (enter file name):')
-            else:
-                embeddings_file = embeddings_files[0]
-            abspath = directory_path + '/' + embeddings_file
-    return abspath
